Adding redundant_metadata test and refactoring gtest suite to allow better code sharing
