+++ tests/cbtest/gtest/Makefile.am
@@ -4,10 +4,11 @@ DEFAULT_INCLUDES += \
 	-I$(top_srcdir)/include \
 	-I$(top_srcdir)/lib/libspl/include
 
-sbin_PROGRAMS = test_uzfs test_zrepl_prot
+sbin_PROGRAMS = test_uzfs test_zfs test_zrepl_prot
 
 test_uzfs_SOURCES = test_uzfs.cc
 test_zrepl_prot_SOURCES = test_zrepl_prot.cc gtest_utils.cc
+test_zfs_SOURCES = test_zfs.cc gtest_utils.cc
 
 test_uzfs_LDADD = \
 	$(top_builddir)/lib/libnvpair/libnvpair.la \
@@ -19,3 +20,4 @@ test_uzfs_LDADD = \
 test_uzfs_CXXFLAGS = -std=c++11
 test_uzfs_LDFLAGS = -pthread -lgtest -lgtest_main
 test_zrepl_prot_LDFLAGS = -pthread -lgtest -lgtest_main
+test_zfs_LDFLAGS = -pthread -lgtest -lgtest_main
